requirements.txt
azure-search-documents~=11.2